Arquivos após ver os videos das resoluções

+++ Mundo 2/ex040.py
@@ -6,13 +6,14 @@ de acordo com a média atingida:
 '''
 
 
-n1 = int(input('Digite a primeira nota: '))
-n2 = int(input('Digite a segunda nota: '))
+n1 = float(input('Digite a primeira nota: '))
+n2 = float(input('Digite a segunda nota: '))
 med = (n1+n2)/2
 print('A sua média foi: {:.2f}'.format(med))
 if med < 5:
     print('Você foi reprovado!')
 elif med >= 5 and med <= 6.9:
+#elif 6.9 > med >= 5:
     print(' Voce está de recuperação!')
 elif med >= 7:
     print('Parabens, voce foi aprovado')
